- Correction of the format price in euro bug - extract function to format price in euro en feetsquared in meter to Utils class
