+++ app/src/androidTest/java/com/haksoftware/p9_da_real_estate_manager/EditFragmentTest.kt
@@ -13,8 +13,6 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import com.haksoftware.p9_da_real_estate_manager.utils.NetworkConnectionRule
-import org.junit.After
-import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
@@ -67,7 +65,16 @@ class EditFragmentTest {
         onView(withId(R.id.button_submit))
             .check(matches(isDisplayed()))
         onView(withId(R.id.button_submit)).perform(click())
+        Thread.sleep(500)
 
+        // Clique sur le premier élément de la liste
+        onView(withId(R.id.recycler_view_real_estates))
+            .perform(
+                RecyclerViewActions.actionOnItemAtPosition<RecyclerView.ViewHolder>(
+                    0,
+                    click()
+                )
+            )
         Thread.sleep(1000)
         val textDescription = onView(withId(R.id.text_description))
             .check(matches(isDisplayed()))
