+++ app/src/main/java/com/haksoftware/p9_da_real_estate_manager/ui/real_estates/RealEstatesFragment.kt
@@ -8,18 +8,15 @@ import android.view.MenuInflater
 import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.OnBackPressedCallback
 import androidx.core.view.MenuHost
 import androidx.core.view.MenuProvider
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Lifecycle
 import androidx.navigation.findNavController
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import androidx.slidingpanelayout.widget.SlidingPaneLayout
 import com.haksoftware.p9_da_real_estate_manager.R
 import com.haksoftware.p9_da_real_estate_manager.data.entity.RealEstateWithDetails
 import com.haksoftware.p9_da_real_estate_manager.databinding.FragmentRealEstatesBinding
@@ -169,45 +166,5 @@ class RealEstatesFragment : Fragment(), MenuProvider, SearchCallback, RealEstate
     override fun onEditRealEstateRequested() {
         val action = RealEstatesFragmentDirections.actionNavRealEstatesToNavEdit(idRealEstateSelected)
         findNavController().navigate(action)
-    }
-}
-
-/**
- * Custom OnBackPressedCallback for handling back button presses when a SlidingPaneLayout is open.
- */
-class RealEstateOnBackPressedCallback(
-    private val slidingPaneLayout: SlidingPaneLayout, private val activity: FragmentActivity
-) : OnBackPressedCallback(slidingPaneLayout.isSlideable && slidingPaneLayout.isOpen), SlidingPaneLayout.PanelSlideListener {
-
-    /**
-     * Handles the back button press to close the SlidingPaneLayout.
-     */
-    override fun handleOnBackPressed() {
-        slidingPaneLayout.closePane()
-    }
-
-    /**
-     * Called when a detail view's position changes.
-     */
-    override fun onPanelSlide(panel: View, slideOffset: Float) {}
-
-    /**
-     * Called when a detail view becomes slid completely open.
-     */
-    override fun onPanelOpened(panel: View) {
-        isEnabled = true
-        activity.invalidateOptionsMenu()
-    }
-
-    /**
-     * Called when a detail view becomes slid completely closed.
-     */
-    override fun onPanelClosed(panel: View) {
-        isEnabled = false
-        activity.invalidateOptionsMenu()
-    }
-
-    init {
-        slidingPaneLayout.addPanelSlideListener(this)
     }
 }
